blogapp/requirements.txt
argon2-cffi==18.3.0
attrs==18.2.0
Automat==0.7.0
bcrypt==3.1.4
captcha==0.3
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
claptcha==0.3.3
colorama==0.4.1
constantly==15.1.0
defusedxml==0.5.0
Django==2.1
django-filter==2.0.0
django-mysql==2.4.1
django-taggit==0.23.0
django-user-email-extension==0.1.11
djangorestframework==3.9.0
Faker==1.0.0
httpie==1.0.2
hyperlink==18.0.0
idna==2.7
incremental==17.5.0
Jinja2==2.7
Markdown==3.0.1
MarkupSafe==1.1.0
mysql-connector-python==8.0.13
mysqlclient==1.3.13
oauthlib==2.1.0
Pillow==5.3.0
protobuf==3.6.1
pycparser==2.19
Pygments==2.3.0
PyHamcrest==1.9.0
PyJWT==1.7.1
python-dateutil==2.7.5
python3-openid==3.1.0
pytz==2018.7
requests==2.21.0
requests-oauthlib==1.0.0
six==1.11.0
social-auth-app-django==3.1.0
social-auth-core==2.0.0
text-unidecode==1.2
Twisted==18.9.0
urllib3==1.24.1
zope.interface==4.6.0
